Bug 732114 - part 6: update SessionTestHelper method names; move to sync.test.helpers package.
