+++ test/src/org/mozilla/android/sync/test/helpers/SessionTestHelper.java
@@ -1,12 +1,10 @@
 /* Any copyright is dedicated to the Public Domain.
    http://creativecommons.org/publicdomain/zero/1.0/ */
 
-package org.mozilla.android.sync.test;
+package org.mozilla.android.sync.test.helpers;
 
 import static junit.framework.Assert.assertNotNull;
 
-import org.mozilla.android.sync.test.helpers.DefaultSessionCreationDelegate;
-import org.mozilla.android.sync.test.helpers.ExpectBeginDelegate;
 import org.mozilla.android.sync.test.helpers.WaitHelper;
 import org.mozilla.gecko.sync.repositories.InvalidSessionTransitionException;
 import org.mozilla.gecko.sync.repositories.Repository;
@@ -17,7 +15,7 @@ import android.util.Log;
 
 public class SessionTestHelper {
 
-  public static RepositorySession prepareRepositorySession(
+  protected static RepositorySession prepareRepositorySession(
       final Context context,
       final boolean begin,
       final Repository repository) {
@@ -73,12 +71,12 @@ public class SessionTestHelper {
 
     return session;
   }
-  
-  public static RepositorySession prepEmptySession(final Context context, final Repository repository) {
-    return prepareRepositorySession(context, true, repository);
-  }
-  
-  public static RepositorySession prepEmptySessionWithoutBegin(Context context, Repository repository) {
+
+  public static RepositorySession createSession(final Context context, final Repository repository) {
     return prepareRepositorySession(context, false, repository);
+  }
+
+  public static RepositorySession createAndBeginSession(Context context, Repository repository) {
+    return prepareRepositorySession(context, true, repository);
   }
 }
